backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
beautifulsoup4>=4.9.0
pdfplumber
sentence-transformers
scikit-learn
numpy
pandas
python-dotenv==1.0.0
aiofiles==23.2.1
azure-cognitiveservices-speech==1.38.0
